+++ .eslintrc
@@ -16,6 +16,7 @@
     "no-use-before-define": "off",
     "react/jsx-no-target-blank": [2, { "allowReferrer": true }],
     "@typescript-eslint/ban-ts-comment": "off",
-    "@typescript-eslint/no-empty-function": "off"
+    "@typescript-eslint/no-empty-function": "off",
+    "@typescript-eslint/explicit-module-boundary-types": "off"
   }
 }
